Split the spelling-gate tests out of the rollout suite (#312)
Rebasing onto main brought in the source-wide spelling work, which grew
`test_typos_rollout.py` to 459 lines — over the 400-line ceiling that
this branch's own pylint enrolment now enforces on `scripts/`. The gate
caught it, so honour it rather than raise the threshold.

The module splits cleanly in two: the script-baseline parsing, the
Makefile spelling-target contract, and the gate's detection behaviour
move to `test_typos_rollout_spelling_gate.py` (225 lines), leaving the
dictionary model, merge, and render tests behind (245 lines). Register
the new module in the `spelling-helper-test` collection alongside the
others.
